# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TradeUpdate API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/trades:
    get:
      operationId: getUserTrades
      tags: [user]
      summary: Get user trade history
      parameters:
        - name: filter
          in: query
          schema:
            type: string
            enum: [all, win, loss, paper, copy]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Trade history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeListResponse"
  /user/signals:
    get:
      operationId: getUserSignals
      tags: [user]
      summary: Get signal log
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Signal log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignalListResponse"
...
  /user/bot/toggle:
    post:
      operationId: toggleBot
      tags: [user]
      summary: Start or stop bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotToggleInput"
      responses:
        "200":
          description: Bot toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotStatus"
...
  /user/chart/candles:
    get:
      operationId: getCandles
      tags: [user]
      summary: Get candlestick data
      parameters:
        - name: timeframe
          in: query
          required: true
          schema:
            type: string
            enum: ["1m", "5m", "15m"]
        - name: count
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Candle data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandleListResponse"
...
  /admin/users/{userId}/dashboard:
    get:
      operationId: adminGetUserDashboard
      tags: [admin]
      summary: View user dashboard as admin
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDashboard"
...
components:
  schemas:
...
    CurrentUser:
      type: object
      required: [id, username, role, status]
      properties:
        id:
          type: string
        username:
          type: string
        role:
          type: string
        status:
          type: string
        email:
          type: string
        tradingProfile:
          type: ["string", "null"]
        strategyId:
          type: ["string", "null"]
        accountBalance:
          type: number
        peakBalance:
          type: number
        dailyStartBalance:
          type: number
        autoCompoundEnabled:
          type: boolean
        adaptiveIntelligenceEnabled:
          type: boolean
        copyTradingEnabled:
          type: boolean
        mustChangePassword:
          type: boolean
        hasDerivToken:
          type: boolean
        tradingMode:
          type: ["string", "null"]
...
    UserDashboard:
      type: object
      required: [user, botStatus, recentTrades, sessionPerformance]
      properties:
        user:
          $ref: "#/components/schemas/CurrentUser"
        botStatus:
          $ref: "#/components/schemas/BotStatus"
        recentTrades:
          type: array
          items:
            $ref: "#/components/schemas/Trade"
        sessionPerformance:
          type: array
          items:
            $ref: "#/components/schemas/SessionSummary"
        dailyPnl:
          type: number
        todayTrades:
          type: integer
        winStreak:
          type: integer
        lossStreak:
          type: integer
    BotStatus:
      type: object
      required: [isRunning, killSwitchActive, tradingMode]
      properties:
        isRunning:
          type: boolean
        killSwitchActive:
          type: boolean
        tradingMode:
          type: string
        todayTrades:
          type: integer
        thisHourTrades:
          type: integer
        consecutiveLosses:
          type: integer
        consecutiveWins:
          type: integer
        dailyPnl:
          type: number
        currentDrawdown:
          type: number
        recoveryModeActive:
          type: boolean
        winStreakCautionActive:
          type: boolean
        pauseReason:
          type: ["string", "null"]
        lastSignalScore:
          type: ["number", "null"]
        lastSignalDirection:
          type: ["string", "null"]
        currentScore:
          type: ["number", "null"]
        openTrade:
          $ref: "#/components/schemas/OpenTrade"
        sessionMultiplier:
          type: number
        adaptiveWeightsActive:
          type: boolean
        rangeContext:
          type: ["string", "null"]
        spikeDetected:
          type: boolean
        consolidationDetected:
          type: boolean
        firstCandleWaiting:
          type: boolean
        strategyCircuitBreakerActive:
          type: boolean
        cooldownSecondsRemaining:
          type: ["integer", "null"]
        scoreBreakdown:
          $ref: "#/components/schemas/ScoreBreakdown"
    ScoreBreakdown:
      type: object
      properties:
        trend:
          type: number
        volatility:
          type: number
        timing:
          type: number
        pullback:
          type: number
        risk:
          type: number
        total:
          type: number
        direction:
          type: string
        trendDirection:
          type: string
        bandTouched:
          type: string
    OpenTrade:
      type: object
      properties:
        id:
          type: string
        direction:
          type: string
        entryPrice:
          type: number
        currentPrice:
          type: number
        stake:
          type: number
        pnl:
          type: number
        stopLoss:
          type: number
        takeProfit1:
          type: number
        takeProfit2:
          type: number
        breakEvenMoved:
          type: boolean
        partialClosed:
          type: boolean
        momentumExtensionActive:
          type: boolean
        openedAt:
          type: integer
    BotToggleInput:
      type: object
      required: [running]
      properties:
        running:
          type: boolean
    Trade:
      type: object
      required: [id, userId, direction, status, openedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        strategyId:
          type: ["string", "null"]
        direction:
          type: string
        entryPrice:
          type: ["number", "null"]
        exitPrice:
          type: ["number", "null"]
        stake:
          type: ["number", "null"]
        pnl:
          type: ["number", "null"]
        pips:
          type: ["number", "null"]
        scoreTotal:
          type: ["number", "null"]
        scoreTrend:
          type: ["number", "null"]
        scoreVolatility:
          type: ["number", "null"]
        scoreTiming:
          type: ["number", "null"]
        scorePullback:
          type: ["number", "null"]
        scoreRisk:
          type: ["number", "null"]
        isCopyTrade:
          type: boolean
        isPaper:
          type: boolean
        tradingMode:
          type: ["string", "null"]
        status:
          type: string
        stopLoss:
          type: ["number", "null"]
        takeProfit1:
          type: ["number", "null"]
        takeProfit2:
          type: ["number", "null"]
        breakEvenMoved:
          type: boolean
        partialClosed:
          type: boolean
        momentumExtensionActive:
          type: boolean
        recoveryModeActive:
          type: boolean
        winStreakCautionActive:
          type: boolean
        sessionName:
          type: ["string", "null"]
        rangeContext:
          type: ["string", "null"]
        openedAt:
          type: integer
        closedAt:
          type: ["integer", "null"]
    TradeListResponse:
      type: object
      required: [trades, total]
      properties:
        trades:
          type: array
          items:
            $ref: "#/components/schemas/Trade"
        total:
          type: integer
    Signal:
      type: object
      required: [id, timestamp, scoreTotal, direction, action]
      properties:
        id:
          type: string
        userId:
          type: string
        strategyId:
          type: ["string", "null"]
        timestamp:
          type: integer
        scoreTotal:
          type: number
        scoreTrend:
          type: number
        scoreVolatility:
          type: number
        scoreTiming:
          type: number
        scorePullback:
          type: number
        scoreRisk:
          type: number
        direction:
          type: string
        action:
          type: string
        rejectionReason:
          type: ["string", "null"]
        ema9:
          type: ["number", "null"]
        ema21:
          type: ["number", "null"]
        rsi:
          type: ["number", "null"]
        sessionName:
          type: ["string", "null"]
        rangeContext:
          type: ["string", "null"]
    SignalListResponse:
      type: object
      required: [signals]
      properties:
        signals:
          type: array
          items:
            $ref: "#/components/schemas/Signal"
...
    SessionSummary:
      type: object
      required: [sessionName, totalTrades, wins, losses, winRate, sessionMultiplier]
      properties:
        sessionName:
          type: string
        totalTrades:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        winRate:
          type: number
        sessionMultiplier:
          type: number
...
    Candle:
      type: object
      required: [time, open, high, low, close, volume]
      properties:
        time:
          type: integer
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number
    CandleListResponse:
      type: object
      required: [candles, timeframe]
      properties:
        candles:
          type: array
          items:
            $ref: "#/components/schemas/Candle"
        timeframe:
          type: string
...