# lib/api-zod/src/generated/types/candleListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TradeUpdate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Candle } from './candle';

export interface CandleListResponse {
  candles: Candle[];
  timeframe: string;
}