# lib/api-zod/src/generated/types/tradeListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TradeUpdate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Trade } from './trade';

export interface TradeListResponse {
  trades: Trade[];
  total: number;
}